third_party/WebKit/LayoutTests/webaudio/osc-square-sweep-snr-expected.txt

Test Square Oscillator with Exponential Sweep

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Exceeded SNR threshold of 87.75 dB
PASS Maximum difference below threshold of 3.63 ulp (16-bits)
PASS Number of differences between actual and expected result is less than 6800
PASS successfullyParsed is true

TEST COMPLETE
